Event-planning-app-main/Event-planning-app-main/requirements.txt
streamlit
google-generativeai
reportlab===4.1.0